Remove action entity
