--- frontend/src/metabase/entities/actions/index.ts
@@ -1,1 +1,0 @@
-export * from "./actions";
