+++ frontend/src/metabase/entities/containers/rtk-query/types/entities.ts
@@ -35,7 +35,6 @@ export type ReloadIntervalSelector<Entity> = (
  * There should be an entry here for every "createEntity" function call.
  */
 export type EntityType =
-  | "actions"
   | "alerts"
   | "bookmarks"
   | "collections"
