+++ frontend/src/metabase/entities/index.ts
@@ -68,7 +68,6 @@
  *   )(BookContainer);
  */
 
-export { Actions as actions } from "./actions/actions";
 export { Collections as collections } from "./collections";
 export { SnippetCollections as snippetCollections } from "./snippet-collections";
 export { Dashboards as dashboards } from "./dashboards";
